feat: MainScreen
